Ensure that the types of the enums values are correct
When the enum package is not available it's necessary to change the type of the _Enum_Type attributes inside a class derived from _Enum to match the type of the containing class (e.g. set the type of NONE inside the class ERR to ERR). Otherwise the behaviour with or without the enum package would be different.