Check and load a functional backend
Though libraries of a given backend can be present on the target
machine, there can be scenarios where a compatible hardware device
is not available. In such cases, backend loader checks if current
selected backend has a device ready for use. If no appropriate
device exists, the next backend in the priority list is checked
for same criteria. This process is repeated until a working
backend is found. In the event that no functional backend is found,
the program throws a runtime error.